requirements.txt
flask
flask-cors
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langchain>=1.0.0
langchain-classic
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
pypdf
docx2txt
unstructured
openpyxl
networkx
pandas
unstructured[xlsx]